Rename LegacyExtension* classes to IndexedExtension*
"Legacy" naming was misleading since these are the default classes.
"Indexed" describes the mechanism: using META-INF/extensions.idx.

Legacy classes are now deprecated aliases for backwards compatibility.

Fixes #643
